+++ java/memory/memory-core/src/main/java/org/apache/arrow/memory/util/hash/ArrowBufHasher.java
@@ -33,7 +33,7 @@ public interface ArrowBufHasher {
    * @param length length of the memory region.
    * @return the hash code.
    */
-  int hashCode(long address, long length);
+  long hashCode(long address, long length);
 
   /**
    * Calculates the hash code for a memory region.
@@ -43,5 +43,5 @@ public interface ArrowBufHasher {
    * @param length length of the memory region.
    * @return the hash code.
    */
-  int hashCode(ArrowBuf buf, long offset, long length);
+  long hashCode(ArrowBuf buf, long offset, long length);
 }
