+++ java/memory/memory-core/src/main/java/org/apache/arrow/memory/util/hash/SimpleHasher.java
@@ -48,7 +48,7 @@ public class SimpleHasher implements ArrowBufHasher {
    * @return the hash code.
    */
   @Override
-  public int hashCode(long address, long length) {
+  public long hashCode(long address, long length) {
     int hashValue = 0;
     int index = 0;
     while (index + 8 <= length) {
@@ -84,7 +84,7 @@ public class SimpleHasher implements ArrowBufHasher {
    * @return the hash code.
    */
   @Override
-  public int hashCode(ArrowBuf buf, long offset, long length) {
+  public long hashCode(ArrowBuf buf, long offset, long length) {
     buf.checkBytes(offset, offset + length);
     return hashCode(buf.memoryAddress() + offset, length);
   }
