+++ java/memory/memory-core/src/test/java/org/apache/arrow/memory/util/TestArrowBufPointer.java
@@ -190,13 +190,13 @@ public class TestArrowBufPointer {
     protected int counter = 0;
 
     @Override
-    public int hashCode(long address, long length) {
+    public long hashCode(long address, long length) {
       counter += 1;
       return SimpleHasher.INSTANCE.hashCode(address, length);
     }
 
     @Override
-    public int hashCode(ArrowBuf buf, long offset, long length) {
+    public long hashCode(ArrowBuf buf, long offset, long length) {
       counter += 1;
       return SimpleHasher.INSTANCE.hashCode(buf, offset, length);
     }
